+++ LtiLibrary.AspNetCore.Tests/Lis/v2/MembershipControllerShould.cs
@@ -3,8 +3,8 @@ using System.IO;
 using System.Net;
 using System.Net.Http;
 using LtiLibrary.AspNetCore.Tests.SimpleHelpers;
+using LtiLibrary.NetCore.Lis.v1;
 using Xunit;
-using LtiLibrary.NetCore.Lti1;
 using LtiLibrary.NetCore.Membership;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
