+++ LtiLibrary.AspNetCore/Lis/v2/GetMembershipDto.cs
@@ -1,5 +1,5 @@
-﻿using LtiLibrary.NetCore.Lis.v2;
-using LtiLibrary.NetCore.Lti1;
+﻿using LtiLibrary.NetCore.Lis.v1;
+using LtiLibrary.NetCore.Lis.v2;
 
 namespace LtiLibrary.AspNetCore.Lis.v2
 {
