+++ LtiLibrary.NetCore/Lis/v2/Membership.cs
@@ -1,4 +1,4 @@
-﻿using LtiLibrary.NetCore.Lti1;
+﻿using LtiLibrary.NetCore.Lis.v1;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
