+++ LtiLibrary.NetCore/Membership/MembershipClient.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using LtiLibrary.NetCore.Common;
 using LtiLibrary.NetCore.Extensions;
+using LtiLibrary.NetCore.Lis.v1;
 using LtiLibrary.NetCore.Lis.v2;
 using LtiLibrary.NetCore.Lti1;
